utils/ferryPilot/requirements.txt
questionary>=2.0.0
pyinstaller>=6.0.0
requests>=2.28.0
python-frontmatter>=1.0.0
tomli-w>=1.0.0
